LPD-100544 Move the portlet export redirect test to the revamp suite
